refactor(cli): extract parsePosition and offerAutoPromote to shared utilities
- Move parsePosition from add.ts to shared.ts for reuse by reorder command
- Extract inline auto-promote logic from add.ts into shared offerAutoPromote
- Remove updateRouteVersion direct import from add.ts (now in shared)
